new top-level interface
