+++ README.md
@@ -1,6 +1,9 @@
 BlackBoxOptim.jl
 ==============
 
+**NOTE!** The top-level API of BlackBoxOptim has just recently changed and we are going through a couple of testing rounds. Things should settle shortly. Sorry for any inconvenience!
+
+
 BlackBoxOptim is a (work-in-progress) global optimization framework for Julia (http://julialang.org/). It supports both multi- and single-objective optimization problems and is focused on (meta-)heuristic/stochastic algorithms (DE, NES, CMA-ES etc) that do NOT require the function being optimized to be differentiable. This is in contrast to more traditional, deterministic algorithms that are often based on gradients/differentiability.
 
 [![Build Status](https://travis-ci.org/robertfeldt/BlackBoxOptim.jl.svg?branch=master)](https://travis-ci.org/robertfeldt/BlackBoxOptim.jl)
@@ -34,19 +37,19 @@ First, we'll load BlackBoxOptim and define the Rosenbrock function (in 2 dimensi
 
 We can now call the `bboptimize` function, specifying the function to be optimized (here: rosenbrock2d) and the range of values allowed for each of the dimensions of the input:
 
-    bboptimize(rosenbrock2d; search_range = (-5.0, 5.0), dimensions = 2)
+    bboptimize(rosenbrock2d; SearchRange = (-5.0, 5.0), NumDimensions = 2)
 
 BlackBoxOptim will default to using an adaptive differential evolution optimizer in this case and use it to try to locate a solution where both elements can be Floats in the range -5.0:5.0. If you wanted a different range of allowed values for the second dimension of the solution you can specify that with a range of allowed values. In this case you do not need to specify the number of dimensions since that is implicit from the number of ranges supplied:
 
-    bboptimize(rosenbrock2d; search_range = [(-5.0, 5.0), (-2.0, 2.0)])
+    bboptimize(rosenbrock2d; SearchRange = [(-5.0, 5.0), (-2.0, 2.0)])
 
 If you want to use a different optimizer that can be specified with the `method` keyword. For example, using the standard differential evolution optimizer DE/rand/1/bin:
 
-    bboptimize(rosenbrock2d; search_range = (-5.0, 5.0), dimensions = 2, method = :de_rand_1_bin)
+    bboptimize(rosenbrock2d; SearchRange = (-5.0, 5.0), NumDimensions = 2, Method = :de_rand_1_bin)
 
 Note that the rosenbrock2d function is quite easy to optimize. Even a random search will come close if we give it more time:
 
-    bboptimize(rosenbrock2d; search_range = (-5.0, 5.0), dimensions = 2, method = :random_search, max_time = 10.0)
+    bboptimize(rosenbrock2d; SearchRange = (-5.0, 5.0), NumDimensions = 2, Method = :random_search, MaxTime = 10.0)
 
 But if we optimize the same rosenbrock function in, say, 30 dimensions that will be very hard for a random searcher while sNES or DE can find good solutions if we give them some time. We can compare optimizers using the `compare_optimizers` function:
 
@@ -54,7 +57,7 @@ But if we optimize the same rosenbrock function in, say, 30 dimensions that will
       return( sum( 100*( x[2:end] - x[1:end-1].^2 ).^2 + ( x[1:end-1] - 1 ).^2 ) )
     end
 
-    res = compare_optimizers(rosenbrock; search_range = (-5.0, 5.0), dimensions = 30, max_time = 5.0);
+    res = compare_optimizers(rosenbrock; SearchRange = (-5.0, 5.0), NumDimensions = 30, MaxTime = 3.0);
 
 You can find more examples of using BlackBoxOptim in [the examples directory](https://github.com/robertfeldt/BlackBoxOptim.jl/tree/master/examples).
 
@@ -74,22 +77,21 @@ The section above described the basic API for the BlackBoxOptim package. We empl
 * `generating_set_search`
 * `probabilistic_descent`
 
-In addition to the `method` keyword, you can alter the behavior of the Optim package by using other keywords:
+In addition to the `Method` parameters, there are many other parameters you can change.
 
-* `max_time`: For how long can the optimization run? Defaults to false which means that number of iterations is the given budget, rather than time.
-* `show_trace`: Should a trace of the optimization be shown on `STDOUT`? Defaults to `false`.
-* `population_size`: How large is the initial population for population-based optimizers? Defaults to `50`.
-* `parameters`: Detailed control over the optimization, through a Dict mapping named parameters to their values.
+* `MaxTime`: For how long can the optimization run? Defaults to false which means that number of iterations is the given budget, rather than time.
+* `ShowTrace`: Should a trace of the optimization be shown on `STDOUT`? Defaults to `false`.
+* `PopulationSize`: How large is the initial population for population-based optimizers? Defaults to `50`.
 
-Most optimizers have specific options that can be specified in the `parameters` dict. Further details TBD.
+You can also have detailed control over the optimization byt giving a Dict mapping named parameters to their values. Most optimizers have specific options that can be specified in the `parameters` dict.
 
 # State of the Library
 
 ## Existing Optimizers
 
 * Natural Evolution Strategies:
-  - Separable NES: `separable_nes()`
-  - Exponential NES: `xnes()`
+  - Separable NES: `separable_nes`
+  - Exponential NES: `xnes`
 * Differential Evolution optimizers, 5 different:
   - Adaptive DE/rand/1/bin: `de_rand_1_bin`
   - Adaptive DE/rand/1/bin with radius limited sampling: `adaptive_de_rand_1_bin_radiuslimited`
