+++ test/problems/common.jl
@@ -6,11 +6,11 @@ function fitness_for_opt(problem, numDimensions, populationSize, numSteps, metho
 
   println("\n$(problem.name), n = $(numdims(problem)), optimizer = $(string(method))")
 
-  best, fitness = bboptimize(problem; method = method, parameters = {
-    :NumDimensions => numDimensions,
-    :PopulationSize => populationSize,
-    :MaxSteps => numSteps
-    })
+  res = bboptimize(problem; Method = method,
+    NumDimensions = numDimensions,
+    PopulationSize = populationSize,
+    MaxSteps = numSteps
+  )
 
-  fitness
+  best_fitness(res)
 end
